requirements.txt
apparmor==4.1.7
attrs==26.1.0
autocommand==2.2.2
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
cryptography==46.0.5
dbus-python==1.4.0
distlib==0.4.0
distro==1.9.0
docutils==0.22.4
filelock==3.25.2
gps==3.27.5
idna==3.11
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
legacy-cgi==2.6.4
LibAppArmor==4.1.7
Mako==1.3.10.dev0
Markdown==3.10.2
MarkupSafe==3.0.3
meson==1.10.2
more-itertools==10.8.0
packaging==26.0
pillow==12.1.1
pkg_resources==81.0.0
platformdirs==4.9.4
psutil==7.2.2
pyasyncore==1.0.5
pycairo==1.29.0
pycparser==3.0
PyGObject==3.54.5
pyinotify==0.9.6
python-discovery==1.2.0
pywal==3.3.0
pyxdg==0.28
Reflector==2023.6.28.0.36.1
requests==2.32.5
setuptools==82.0.1
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
urwid==3.0.5
virtualenv==21.2.0
wcwidth==0.6.0
wheel==0.46.3
streamlit>=1.33
langchain-core>=1.2.23
langchain-community>=0.4.1
langchain-google-genai>=4.2.1
langchain-huggingface>=1.2.1
langchain-text-splitters>=1.1.1
faiss-cpu>=1.13.2
sentence-transformers>=5.3.0
pypdf>=4.0.0
python-dotenv>=1.0.0